Improve run comparison tools
Change the methods in mafRunComparison so the resulting comparison
dataframe is an attribute of the class. Added a new method to
normalize the comparison dataframe based on user provided run.
Added a new method to mafRunComparison to plot the normalized
comparison dataframe as a function of run. A set of new
functions are defined in mafMultiRunPlots.py to overplot or
create subplots of already computed metrics.